Change test package structure
