+++ src/test/java/org/ballerinalang/completion/BallerinaCompletionTestBase.java
@@ -1,0 +1,53 @@
+/*
+ *  Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
+package org.ballerinalang.completion;
+
+import com.intellij.codeInsight.completion.CompletionType;
+import com.intellij.psi.PsiFile;
+import com.intellij.testFramework.fixtures.LightPlatformCodeInsightFixtureTestCase;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+
+import java.util.List;
+
+public abstract class BallerinaCompletionTestBase extends LightPlatformCodeInsightFixtureTestCase {
+
+    void doTest(String fileContent, String... expectedLookups) {
+        if (fileContent != null) {
+            myFixture.configureByText("test.bal", fileContent);
+        }
+        myFixture.complete(CompletionType.BASIC, 1);
+        List<String> lookupElementStrings = myFixture.getLookupElementStrings();
+        assertNotNull(lookupElementStrings);
+        assertSameElements(lookupElementStrings, expectedLookups);
+    }
+
+    void doCheckResult(@NotNull String relativePath, @NotNull String before, String after,
+                       @Nullable Character c, String... expectedLookups) {
+        PsiFile testFile = myFixture.addFileToProject(relativePath, before);
+        myFixture.configureFromExistingVirtualFile(testFile.getVirtualFile());
+        myFixture.completeBasic();
+        if (c != null) {
+            myFixture.type(c);
+        }
+        if (after != null) {
+            myFixture.checkResult(after);
+        } else {
+            doTest(null, expectedLookups);
+        }
+    }
+}
