--- src/test/java/psi/vars/BallerinaResolveGlobalVariableTest.java
@@ -1,67 +1,0 @@
-/*
- *  Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *  http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-
-package psi.vars;
-
-import psi.BallerinaResolveTestBase;
-
-import java.io.IOException;
-
-public class BallerinaResolveGlobalVariableTest extends BallerinaResolveTestBase {
-
-    private String globalVariable = "int /*def*/a;";
-
-    @Override
-    protected String getTestDataPath() {
-        return "src/test/resources/testData/psi/resolve/vars/global";
-    }
-
-    public void testGlobalVariableInSameFileInConnector() {
-        doFileTest();
-    }
-
-    public void testGlobalVariableInSameFileInFunction() {
-        doFileTest();
-    }
-
-    public void testGlobalVariableInSameFileInService() {
-        doFileTest();
-    }
-
-    public void testGlobalVariableInDifferentFileInConnector() throws IOException {
-        doFileTest(globalVariable);
-    }
-
-    public void testGlobalVariableInDifferentFileInFunction() throws IOException {
-        doFileTest(globalVariable);
-    }
-
-    public void testGlobalVariableInDifferentFileInService() throws IOException {
-        doFileTest(globalVariable);
-    }
-
-    public void testGlobalVariableInDifferentPackageInConnector() throws IOException {
-        doFileTest(globalVariable, "org/test/test.bal");
-    }
-
-    public void testGlobalVariableInDifferentPackageInFunction() throws IOException {
-        doFileTest(globalVariable, "org/test/test.bal");
-    }
-
-    public void testGlobalVariableInDifferentPackageInService() throws IOException {
-        doFileTest(globalVariable, "org/test/test.bal");
-    }
-}
